# Improve performance with per-type overloads for `PyObject.From`