--- a/src/CSnakes.Runtime/Python/PyObject.cs
+++ b/src/CSnakes.Runtime/Python/PyObject.cs
@@ -541,39 +541,159 @@ public partial class PyObject : SafeHandle, ICloneable
     public T BareImportAs<T, TImporter>() where TImporter : IPyObjectImporter<T> =>
         TImporter.BareImport(this);
 
-    public static PyObject From<T>(T value)
+    public static PyObject From(bool? value) =>
+        value switch { null => None, { } some => From(some) };
+
+    public static PyObject From(bool value) =>
+        value ? True : False;
+
+    public static PyObject From(long? value) =>
+        value switch { null => None, { } n => From(n) };
+
+    public static PyObject From(long value)
+    {
+        switch (value)
+        {
+            case 0: return Zero;
+            case 1: return One;
+            case -1: return NegativeOne;
+            case var n:
+                using (GIL.Acquire())
+                    return Create(CPythonAPI.PyLong_FromLongLong(n));
+        }
+    }
+
+    public static PyObject From(double? value) =>
+        value switch { null => None, { } n => From(n) };
+
+    public static PyObject From(double value)
+    {
+        switch (value)
+        {
+            case 0: return Zero;
+            case 1: return One;
+            case -1: return NegativeOne;
+            case var n:
+                using (GIL.Acquire())
+                    return Create(CPythonAPI.PyFloat_FromDouble(n));
+        }
+    }
+
+    public static PyObject From(BigInteger? value) =>
+        value switch { null => None, { } some => From(some) };
+
+    public static PyObject From(BigInteger value)
+    {
+        switch (value)
+        {
+            case { IsZero: true }: return Zero;
+            case { IsOne: true }: return One;
+            case var n when n == -1: return NegativeOne;
+            case var n:
+                using (GIL.Acquire())
+                    return PyObjectTypeConverter.ConvertFromBigInteger(n);
+        }
+    }
+
+    public static PyObject From(string? value)
     {
         switch (value)
         {
             case null: return None;
-            case true: return True;
-            case false: return False;
-            case 0 or 0L or BigInteger { IsZero: true }: return Zero;
-            case 1 or 1L or BigInteger { IsOne: true }: return One;
-            case -1 or -1L:
-            case BigInteger n when n == -1: return NegativeOne;
-            default:
-            {
+            case var some:
                 using (GIL.Acquire())
-                {
-                    return value switch
-                    {
-                        ICloneable pyObject => pyObject.Clone(),
-                        int i => Create(CPythonAPI.PyLong_FromLong(new(i))),
-                        long l => Create(CPythonAPI.PyLong_FromLongLong(l)),
-                        double d => Create(CPythonAPI.PyFloat_FromDouble(d)),
-                        float f => Create(CPythonAPI.PyFloat_FromDouble((double)f)),
-                        string s => Create(CPythonAPI.AsPyUnicodeObject(s)),
-                        byte[] bytes => PyObject.Create(CPythonAPI.PyBytes_FromByteSpan(bytes.AsSpan())),
-                        IDictionary dictionary => PyObjectTypeConverter.ConvertFromDictionary(dictionary),
-                        ITuple t => PyObjectTypeConverter.ConvertFromTuple(t),
-                        ICollection l => PyObjectTypeConverter.ConvertFromList(l),
-                        IEnumerable e => PyObjectTypeConverter.ConvertFromList(e),
-                        BigInteger b => PyObjectTypeConverter.ConvertFromBigInteger(b),
-                        _ => throw new InvalidCastException($"Cannot convert {value} to PyObject"),
-                    };
-                }
-            }
+                    return Create(CPythonAPI.AsPyUnicodeObject(some));
+        }
+    }
+
+    public static PyObject From(byte[]? value)
+    {
+        switch (value)
+        {
+            case null: return None;
+            case var some:
+                using (GIL.Acquire())
+                    return Create(CPythonAPI.PyBytes_FromByteSpan(some.AsSpan()));
+        }
+    }
+
+    public static PyObject From(IDictionary? value)
+    {
+        switch (value)
+        {
+            case null: return None;
+            case var some:
+                using (GIL.Acquire())
+                    return PyObjectTypeConverter.ConvertFromDictionary(some);
+        }
+    }
+
+    public static PyObject From(ITuple? value)
+    {
+        switch (value)
+        {
+            case null: return None;
+            case var some:
+                using (GIL.Acquire())
+                    return PyObjectTypeConverter.ConvertFromTuple(some);
+        }
+    }
+
+    internal static PyObject From(ICloneable value)
+    {
+        using (GIL.Acquire())
+            return value.Clone();
+    }
+
+    public static PyObject From(ICollection? value)
+    {
+        switch (value)
+        {
+            case null:
+                return None;
+            case IDictionary dictionary:
+                return From(dictionary);
+            case var collection:
+                using (GIL.Acquire())
+                    return PyObjectTypeConverter.ConvertFromList(collection);
+        }
+    }
+
+    public static PyObject From(IEnumerable? value)
+    {
+        switch (value)
+        {
+            case null:
+                return None;
+            case ICloneable cloneable:
+                return From(cloneable);
+            case IDictionary dictionary:
+                return From(dictionary);
+            case var enumerable:
+                using (GIL.Acquire())
+                    return PyObjectTypeConverter.ConvertFromList(enumerable);
+        }
+    }
+
+    public static PyObject From(object? value)
+    {
+        switch (value)
+        {
+            case null: return None;
+            case bool b: return From(b);
+            case int n: long l = n; return From(l);
+            case long n: return From(n);
+            case BigInteger n: return From(n);
+            case float n: double d = n; return From(d);
+            case double n: return From(n);
+            case string s: return From(s);
+            case byte[] bytes: return From(bytes);
+            case ICloneable cloneable: return From(cloneable);
+            case IDictionary dictionary: return From(dictionary);
+            case ITuple tuple: return From(tuple);
+            case ICollection collection: return From(collection);
+            case IEnumerable enumerable: return From(enumerable);
+            default: throw new InvalidCastException($"Cannot convert {value} to PyObject");
         }
     }