@tonybaloney 2025-05-05
Is there any chance of this introducing a new ambiguous overload compiler error for existing code?