@atifaziz 2025-05-05
> Is there any chance of this introducing a new ambiguous overload compiler error for existing code?

I don't believe so, as far as I can tell mentally and tests having the necessary coverage for the intended cases. Ambiguity errors would be caught at compile-time so you wouldn't even need to run the tests. That said, two this:

1. this is a source-compatible but binary-breaking change, though not sure we've been caring about the latter since there's been a few of those.
2. Some `PyObject.From` may need more [coverage](https://github.com/user-attachments/files/20047987/coverage.zip) now that there's a more explicit split. For example, while `PyObject.From(bool)` is covered, `PyObject.From(bool?)` isn't anymore. It may make sense to add them as part of this PR.